+++ src/java/net/sf/jabref/FieldComparator.java
@@ -151,12 +151,12 @@ public class FieldComparator implements Comparator<BibtexEntry> {
                 // The first one was parseable, but not the second one.
                 // This means we consider one < two
                 f1 = i1;
-                f2 = i1.intValue() + 1;
+                f2 = i1 + 1;
             } else if (i2 != null) {
                 // The second one was parseable, but not the first one.
                 // This means we consider one > two
                 f2 = i2;
-                f1 = i2.intValue() + 1;
+                f1 = i2 + 1;
             }
             // Else none of them were parseable, and we can fall back on comparing strings.    
         }
